mengatur controller blueprint
