+++ util/config.py
@@ -1,6 +1,7 @@
 from datetime import timedelta
 import os
 from flask_sqlalchemy import SQLAlchemy
+from google.cloud.sql.connector import Connector, IPTypes
 
 basedir = os.path.abspath(os.path.dirname(__file__))
 
@@ -29,6 +30,20 @@ class Config(object):
     TIMEZONE = "Asia/Jakarta"
     JWT_ACCESS_TOKEN_EXPIRES = timedelta(minutes=30)
 
+    JWT_SECRET_KEY = "283674515990178098796700912839185640515"
     JWT_TOKEN_LOCATION = ["headers"]
     JWT_HEADER_NAME = "Authorization"
     JWT_HEADER_TYPE = "Bearer"
+
+
+def getconn():
+    with Connector() as connector:
+        conn = connector.connect(
+            "testing-flask-api:asia-southeast2:flask-api-db-instance",  # Cloud SQL Instance Connection Name
+            "pymysql",
+            user="riski-db",
+            password="riski123",
+            db="tani_aid",
+            ip_type=IPTypes.PUBLIC,  # IPTypes.PRIVATE for private IP
+        )
+        return conn
